Rename Slugabble to Sluggable, fix typo
